full stable admin notification

+++ RiyadhEmirates_BackEnd/Emirates.Core/Application/Services/ServieNotifications/ServieNotificationService.cs
@@ -23,10 +23,15 @@ namespace Emirates.Core.Application.Services.ServieNotifications
 
         public IApiResponse GetById(int id)
         {
-            var servieNotification = _emiratesUnitOfWork.ServieNotifications.FirstOrDefault(l => l.Id.Equals(id), x => x.Service, x => x.Stage);
+            var servieNotification = _emiratesUnitOfWork.ServieNotifications.FirstOrDefault(l => l.Id.Equals(id), x => x.Service, x => x.Stage, x => x.ServieNotificationLogs);
             if (servieNotification == null)
                 throw new NotFoundException(typeof(ServieNotification).Name);
             return GetResponse(data: _mapper.Map<GetServieNotificationDetailsDto>(servieNotification));
+        }
+        public IApiResponse GetServieNotificationLog(int id)
+        {
+            var servieNotifications = _emiratesUnitOfWork.ServieNotificationLogs.Where(l => l.ServieNotificationId == id);
+            return GetResponse(data: _mapper.Map<List<GetServieNotificationLogListDto>>(servieNotifications));
         }
         public IApiResponse GetAll(SearchModel searchModel)
         {
@@ -52,11 +57,15 @@ namespace Emirates.Core.Application.Services.ServieNotifications
             if (_emiratesUnitOfWork.ServieNotifications.Where(x => x.ServiceId == createModel.ServiceId && x.StageId == createModel.StageId &&x.IsSMS == createModel.IsSMS && x.IsEmail == createModel.IsEmail).Any())
                 throw new BusinessException("تم اضافة الرسالة مسبقا");
 
-            var addedModel = _emiratesUnitOfWork.ServieNotifications.Add(_mapper.Map<ServieNotification>(createModel));
-            addedModel.ServieNotificationLogs.Add(new ServieNotificationLog
+            var addedModel = _mapper.Map<ServieNotification>(createModel);
+            _emiratesUnitOfWork.ServieNotifications.Add(addedModel);
+            _emiratesUnitOfWork.Complete();
+            ServieNotificationLog servieNotificationLog = new ServieNotificationLog
             {
+                ServieNotificationId = addedModel.Id,
                 Message = createModel.Message
-            });
+            };
+            _emiratesUnitOfWork.ServieNotificationLogs.Add(servieNotificationLog);
             _emiratesUnitOfWork.Complete();
             return GetResponse(message: CustumMessages.SaveSuccess(), data: addedModel.Id);
         }
@@ -66,24 +75,28 @@ namespace Emirates.Core.Application.Services.ServieNotifications
             if (servieNotification == null)
                 throw new NotFoundException(typeof(ServieNotification).Name);
 
+            if (_emiratesUnitOfWork.ServieNotifications.Where(x => x.Id != updateModel.Id && x.ServiceId == updateModel.ServiceId && x.StageId == updateModel.StageId &&x.IsSMS == updateModel.IsSMS && x.IsEmail == updateModel.IsEmail).Any())
+                throw new BusinessException("يوجد رسالة مضافة مسبقا على نفس البيانات");
+
+            updateModel.IsSMS = servieNotification.IsSMS;
+            updateModel.IsEmail = servieNotification.IsEmail;
+            updateModel.IsDefault = servieNotification.IsDefault;
+            _emiratesUnitOfWork.ServieNotifications.Update(servieNotification, _mapper.Map<ServieNotification>(updateModel));
+
             if (updateModel.Message != servieNotification.ServieNotificationLogs.FirstOrDefault()?.Message)
             {
-                if (_emiratesUnitOfWork.ServieNotifications.Where(x => x.Id != updateModel.Id && x.ServiceId == updateModel.ServiceId && x.StageId == updateModel.StageId &&x.IsSMS == updateModel.IsSMS && x.IsEmail == updateModel.IsEmail).Any())
-                    throw new BusinessException("يوجد رسالة مضافة مسبقا على نفس البيانات");
-
                 var logs = _emiratesUnitOfWork.ServieNotificationLogs.Where(x => x.ServieNotificationId == updateModel.Id && x.EndDate == null).ToList();
                 foreach (var item in logs)
                 {
                     item.EndDate = DateTime.Now;
                 }
-                _emiratesUnitOfWork.ServieNotifications.Update(servieNotification, _mapper.Map<ServieNotification>(updateModel));
                 _emiratesUnitOfWork.ServieNotificationLogs.Add(new ServieNotificationLog
                 {
                     ServieNotificationId = updateModel.Id,
                     Message = updateModel.Message
                 });
-                _emiratesUnitOfWork.Complete();
             }
+            _emiratesUnitOfWork.Complete();
             return GetResponse(message: CustumMessages.UpdateSuccess(), data: updateModel.Id);
         }
         public IApiResponse ChangeStatus(int id)
@@ -98,10 +111,13 @@ namespace Emirates.Core.Application.Services.ServieNotifications
         }
         public IApiResponse Delete(int id)
         {
-            var servieNotification = _emiratesUnitOfWork.ServieNotifications.FirstOrDefault(n => n.Id == id);
+            var servieNotification = _emiratesUnitOfWork.ServieNotifications.FirstOrDefault(n => n.Id == id, x => x.ServieNotificationLogs);
             if (servieNotification == null)
                 throw new NotFoundException(typeof(ServieNotification).Name);
+            if(servieNotification.IsDefault)
+                throw new BusinessException("لا يمكنك حذف هذا العنصر لكونه العنصر الافتراضي");
 
+            _emiratesUnitOfWork.ServieNotificationLogs.RemoveRange(servieNotification.ServieNotificationLogs);
             _emiratesUnitOfWork.ServieNotifications.Remove(servieNotification);
             _emiratesUnitOfWork.Complete();
             return GetResponse(message: CustumMessages.DeleteSuccess());
